Update and remove assets that shouldn't be included
LiteralExemptions is something that is no longer being used, so we should remove it from the source/package. I moved the rest of the ruleset files that we want to have into the rulesets directory.

I moved the .targets file and additional files into a configuration directory that are used to configure the analyzers to run properly. This can then get packaged into a content directory or something else.