Added remaining rulesets, cleaned up stale configuration and analyzers and added documentation around banned methods (#8)
* Create first pass of analyzers and tests that build and all unit tests pass

* Collapse Analyzers.Naming into Analyzers to consolidate. Moved tests and analyzer code, consolidated namespaces and deleted existing projects.

* Rename assemblies and folders to match. Update namespaces to get all code compiling and tests passing

* Renamed Everything.sln to NationalInstruments.Analyzers.sln.
Renamed test folder to tests to match dotnet github project

* Delete packages.config as it wasn't necessary and cleaned up nuget.config

* Add content from the analyzer wrapper package user branch. Configured these projects to use NI.ruleset and fixed warnings.
NOTE: There's a chicken-and-egg problem because we want to also consume the analyzers that we build here as a part of the ruleset, but we need to get an initial pass working first.
Once we build a package of our analzyers, we will start consuming that packge on newer versions of itself to ensure we're also eating our own dog food.

* Removing files not required to make release builds pass

* Created documentation on using banned methods

Created a .md file to explain how to use the Banned Methods analyzer rule

* Update and remove assets that shouldn't be included

LiteralExemptions is something that is no longer being used, so we should remove it from the source/package. I moved the rest of the ruleset files that we want to have into the rulesets directory.

I moved the .targets file and additional files into a configuration directory that are used to configure the analyzers to run properly. This can then get packaged into a content directory or something else.

* Added back BannedMethods additional files

Added additional file for banned methods to ensure it's picked up by the analyzer

* Fix tabs vs. whitespacing issue.

* Update .editorconfig to have indent size for XML config files

Update .editorconfig to have indent size for XML config files

* Rename properties for NI prefixing

Renamed NI_IsTestProject and NI_IsTestUtilitiesProject

* Prefix properties with NI if we defined them

Prefix properties with NI if we defined them

* Update documentation on banned methods

Update documentation on banned methods

* Update banned method documentation

Removed #2 from the conditions.

* Address review comments, add link and rename ruleset

Added a link to the global exceptions and updated the <RuleSet> to be match the file.

* Remove stale comments

Remove stale comments