Add script to publish "clean" conventions document to github pages (#13)
I leveraged the python-styleguide code that was written to simplify links and clean up things in the conventions document and that will be what you see when you view documentation through github pages. After speaking to Josh, the code can be leveraged as-is since our conventions file name matches theirs and the code in index.md is just looking for that name.