Update C.1.1 to clarify company name should be NationalInstruments (#72)
For a new product namespace we had some discussion about whether we should be using "NI" or "NationalInstruments" after the company name change. There is already an analyzer rule that enforces using "NationalInstruments" so the choice is clear, but the conventions do not state this explicitly. This updates the README.md to clarify this case.